Add Rating System Final Version 1.7

+++ src/main/java/team06/platform/dao/IAccountDao.java
@@ -10,7 +10,7 @@ public interface IAccountDao {
 
     void insertAccount(Long userId, String userName);
 
-    Integer queryBalance(Long userId);
+    Double queryBalance(Long userId);
 
     boolean updateBalance(Account account);
